+++ src/main/java/model/repositories/PasajeroRepository.java
@@ -4,5 +4,5 @@ import model.models.Pasajero;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface PasajeroRepository extends JpaRepository<Pasajero, Integer> {
+public interface PasajeroRepository extends JpaRepository<Pasajero, Long> {
 }
