+++ src/main/java/model/repositories/VueloRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Date;
 import java.util.List;
 
-public interface VueloRepository extends JpaRepository<Vuelo, Integer> {
+public interface VueloRepository extends JpaRepository<Vuelo, Long> {
 }
